+++ spec/hashie/extensions/mash/safe_assignment_spec.rb
@@ -1,0 +1,17 @@
+require 'spec_helper'
+
+describe Hashie::Extensions::Mash::SafeAssignment do
+  class MashWithSafeAssignment < Hashie::Mash
+    include Hashie::Extensions::Mash::SafeAssignment
+  end
+
+  context 'when included in Mash' do
+    subject { MashWithSafeAssignment.new }
+
+    context 'when attempting to override a method' do
+      it 'raises an error' do
+        expect { subject.zip = 'Test' }.to raise_error(ArgumentError)
+      end
+    end
+  end
+end